Updated the U-net archecture
After playing around with layers, settings, custom loss, and custom optimizers. I have achieved my best results with this layout so far. I also changed the image, which seemed to over come the diffculties I was having on the orginal photo. I believe that the orginal photo was too complicated for the network to train on due to the sand grains colors and the spotted details/camo on the fish. More images will be test to learn more.

The amazing thing is that this neural network was able to train on a heavily augmented image set based on the new image (#6). The augumentation include random; rotations, flips, and shifts. This is the first time I have achieved an image very similar to the ground truth with high levels of augemtation. Past attempts were very blurry.

I will continue make tweaks on the model, training, and images for Stage1. When I am confident that the model can handle a dataset of different images I will post Stage 2.